% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{Normalise data}
\usage{
normalise(x)
}
\arguments{
\item{x}{a vector of numeric values}
}
\value{
a vector of normalised values
}
\description{
The function is built to normalise a vector of numbers.
}
\examples{
normalise(c(1:5))
}
