% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseDf.R
\name{normaliseDf}
\alias{normaliseDf}
\title{Normalise dataframe}
\usage{
normaliseDf(dataframe, applicableColumns = c())
}
\arguments{
\item{dataframe}{a dataframe to be normalised}

\item{applicableColumns}{a vector of character of dataframe names to be normalised in the dataframe}
}
\value{
a dataframe with normalised values
}
\description{
This function will normalise the dataframe entered.
If only certain columns need to be normalised, these can be specified.
The order of the columns is kept the same
}
\examples{
dataframe <- data.frame(first = c(1:3), second = c(4:6), third = c(7:9))
normaliseDf(matt)
normaliseDf(matt, c("first", "third"))
}
