% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNaRows.R
\name{removeNaRows}
\alias{removeNaRows}
\title{Remove rows with Na Values from dataframe}
\usage{
removeNaRows(dataframe, applicableColumns = c())
}
\arguments{
\item{dataframe}{a dataframe}

\item{applicableColumns}{a vector of dataframe names which should be checked, ignore others}
}
\value{
The dataframe with rows which have NA values removed
}
\description{
The function will remove all rows which contain na values.
Specific columns can be cleaned whilst ignoring the other columns
}
\examples{
removeNaRows(data)
removeNaRows(read.csv("train.csv"))
removeNaRows(data, c("A", "B", "C"))
}
